Files open with obj host and no saves in localStorage, various files open with same name at same time, clean file information when no more needed.
getFile() changed for open() and require object that want to open the file, also not save file in local storage.
Can open several files at same time withoout save them, and all are diferent files, when one is saved impact in all files that are open in
the same tab.
Gsfile know the objects that open it, when all of them close GsFile fire fileclosed event to allow deleted from memory.

.-